# cedar-monitor-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe monitor server reports on the rest of CEDAR: how deep the work queues are, how much each\nstore holds, whether a given server considers itself healthy, and what artifact a given string\nrefers to.\n\nEverything here is read-only and needs the monitor read permission. The counts are meant to be\nread against each other: the workspace graph is authoritative, and the search index disagreeing\nwith it is how a half-finished reindex shows itself.\n",
    "title" : "CEDAR Monitor Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/logs/explorer/requests" : {
      "get" : {
        "description" : "One row per request the raw request log still holds, which covers roughly the last thirty days. Each row carries the caller, the handler that served it, the status and the handler duration in nanoseconds. Requires the monitor read permission.",
        "operationId" : "requests",
        "parameters" : [ {
          "description" : "Free-text filter, matched as a substring against the path, the user identifier, the handler class and the global request identifier. Omit to match every row.",
          "in" : "query",
          "name" : "q",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Keep only requests whose handler took at least this many milliseconds. Zero and negative values are treated as no filter.",
          "in" : "query",
          "name" : "minDurationMs",
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "description" : "How many rows to return. Defaults to 100 and is capped at 500.",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RequestRow"
                  }
                }
              }
            },
            "description" : "Matching request rows, newest first"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "List recent HTTP requests, newest first",
        "tags" : [ "Log explorer" ]
      }
    },
...
    "/logs/usage/summary" : {
      "get" : {
        "description" : "The totals over the range beside the hourly series they were summed from. Both come from the hourly rollups, so the newest hour appears only once the aggregator has closed it. Requires the monitor read permission.",
        "operationId" : "summary",
        "parameters" : [ {
          "description" : "Inclusive lower bound, as an ISO-8601 instant in UTC. Defaults to seven days before the upper bound.",
          "in" : "query",
          "name" : "from",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Exclusive upper bound, as an ISO-8601 instant in UTC. Defaults to now.",
          "in" : "query",
          "name" : "to",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LogUsageSummary"
                }
              }
            },
            "description" : "Totals and the hourly series behind them"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read, or a bound was not an ISO-8601 instant"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get request volume and latency totals for a time range",
        "tags" : [ "Log usage" ]
      }
    },
...
    "/server-report/{server}/build" : {
      "get" : {
        "description" : "The version the named server's environment declares, together with the path and modification time of the artifact its JVM actually loaded, its uptime and its host.",
        "operationId" : "build",
        "parameters" : [ {
          "description" : "CEDAR server name, as the configuration spells it.",
          "in" : "path",
          "name" : "server",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServerBuildReport"
                }
              }
            },
            "description" : "The named server's build report"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "No CEDAR server answers to this name"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The named server could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get a server's version and build report",
        "tags" : [ "Server report" ]
      }
    },
    "/server-report/{server}/configuration" : {
      "get" : {
        "description" : "cedar-main.yml as the named server resolved it, with secrets masked. Placeholders that could not be resolved are left as the literal ${NAME}.",
        "operationId" : "configuration",
        "parameters" : [ {
          "description" : "CEDAR server name, as the configuration spells it.",
          "in" : "path",
          "name" : "server",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServerConfigurationReport"
                }
              }
            },
            "description" : "The named server's resolved configuration"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "No CEDAR server answers to this name"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The named server could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get a server's resolved configuration",
        "tags" : [ "Server report" ]
      }
    },
    "/server-report/{server}/environment" : {
      "get" : {
        "description" : "The environment variables the named server resolved its configuration from, with secret values masked and undeclared values withheld. Read one server per call; the monitoring UI reads all of them in parallel to build the drift matrix.",
        "operationId" : "environment",
        "parameters" : [ {
          "description" : "CEDAR server name, as the configuration spells it. Examples: artifact, terminology, resource.",
          "in" : "path",
          "name" : "server",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServerEnvironmentReport"
                }
              }
            },
            "description" : "The named server's environment report"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "No CEDAR server answers to this name"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The named server could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get a server's environment variable report",
        "tags" : [ "Server report" ]
      }
    },
    "/server-report/{server}/insight" : {
      "get" : {
        "description" : "Heap and non-heap memory, thread counts, garbage collection counts and host load for the named server. Thread stack traces are deliberately not included: the full dump runs to roughly half a megabyte per server and is served by the server's own /insight/thread-details.",
        "operationId" : "insight",
        "parameters" : [ {
          "description" : "CEDAR server name, as the configuration spells it.",
          "in" : "path",
          "name" : "server",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServerJvmReport"
                }
              }
            },
            "description" : "The named server's JVM report"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "No CEDAR server answers to this name"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The named server could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get a server's JVM report",
        "tags" : [ "Server report" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CedarError" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "A CEDAR error response. Diagnostic fields are populated only when they are relevant to the failure.",
        "properties" : {
          "entities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named domain entities associated with exception-mapped errors.",
              "nullable" : true
            },
            "description" : "Named domain entities associated with exception-mapped errors.",
            "nullable" : true
          },
          "errorId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Correlation identifier for an internal error recorded in server logs.",
            "nullable" : true
          },
          "errorKey" : {
            "type" : "string",
            "description" : "Stable machine-readable error identifier.",
            "enum" : [ "templateElementNotCreated", "templateElementNotFound", "templateElementNotDeleted", "templateElementNotUpdated", "templateElementsNotListed", "templateFieldNotCreated", "templateFieldNotFound", "templateFieldNotDeleted", "templateFieldNotUpdated", "templateFieldsNotListed", "templateNotCreated", "templateNotFound", "templateNotDeleted", "templateNotUpdated", "templatesNotListed", "templateInstanceNotCreated", "templateInstanceNotFound", "templateInstanceNotDeleted", "templateInstanceNotUpdated", "templateInstancesNotListed", "noReadAccessToResource", "noWriteAccessToResource", "noReadAccessToFolder", "noWriteAccessToFolder", "noReadAccessToTemplate", "noWriteAccessToTemplate", "noReadAccessToTemplateElement", "noWriteAccessToTemplateElement", "noReadAccessToTemplateField", "noWriteAccessToTemplateField", "noReadAccessToTemplateInstance", "noWriteAccessToTemplateInstance", "noReadAccessToArtifact", "noWriteAccessToArtifact", "noReadAccessToCategory", "noWriteAccessToCategory", "folderNotFound", "artifactNotFound", "artifactPreconditionRequired", "artifactHasMovedOn", "verbatimWriteRefused", "nodeNotFound", "unknownResourceType", "missingParameter", "invalidInput", "sourceFolderNotFound", "sourceResourceNotFound", "targetFolderNotFound", "nodeNotMoved", "groupNotFound", "userNotFound", "uniqueConstraintCollision", "invalidData", "notAuthorized", "permissionMissing", "tokenInvalid", "userInfoLoadByTokenFailed", "userInfoLoadByApiKeyFailed", "cedarUserNotFound", "authorizationNotFound", "permissionNotOwned", "authorizationTypeUnknown", "apiKeyNotFound", "tokenMissing", "tokenExpired", "parentFolderNotSpecified", "parentFolderSpecifiedTwice", "pathNotNormalized", "parentFolderNotFound", "updateInvalidFolderName", "createInvalidFolderName", "nodeAlreadyPresent", "folderNotCreated", "missingNameAndDescription", "missingData", "folderNotDeleted", "artifactNotDeleted", "resourceNotCreated", "invalidResourceType", "invalidArtifactType", "readOtherProfileForbidden", "updateOtherProfileForbidden", "folderCanNotBeDeleted", "folderCanNotBeChanged", "groupAlreadyPresent", "groupCanBeModifiedOnlyByGroupAdmin", "groupCanBeDeletedOnlyByGroupAdmin", "groupMembersCanBeReadOnlyByGroupAdmin", "groupRequiresAdministrator", "groupUsersNotUpdated", "specialGroupCanNotBeDeleted", "folderPermissionsCanNotBeChanged", "unknownInstanceOutputFormat", "folderCopyNotAllowed", "methodNotImplemented", "upstreamServerError", "nothingToDo", "parentCategoryNotFound", "categoryAlreadyPresent", "categoryNotFound", "categoryCanNotBeDeleted", "rootCategoryCanNotBeDeleted", "unableToAttachCategory", "unableToDetachCategory", "noCategoriesWereAttached", "malformedJsonRequestBody", "malformedSearchTerm", "publishedArtifactCanNotBeChanged", "publishedArtifactCanNotBeDeleted", "versioningOnlyOnLatest", "versioningOnlyByOwner", "nonVersionedArtifactType", "createDraftOnlyFromPublished", "publishOnlyDraft", "draftNotCreated", "contentNotValid", "doiNotSupportedByResourceType", "doiCanNotBeSetForEmptyAtId", "doiCanNotBeAltered", "doiCanNotBeSet", "doiAlreadyExists", "dataCiteDOIDisabled", "resourceNotFound", "internalError" ],
            "nullable" : true
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Alias of message retained for compatibility.",
            "nullable" : true
          },
          "errorReasonKey" : {
            "type" : "string",
            "description" : "More specific machine-readable reason for the error.",
            "enum" : [ "nonEmptyFolder", "nonEmptyCategory", "userHomeFolder", "systemFolder", "templateReferencedInInstances", "validationError", "continuationExpired" ],
            "nullable" : true
          },
          "errorType" : {
            "type" : "string",
            "description" : "Broad category of the error.",
            "enum" : [ "notFound", "invalidArgument", "authorization", "permission", "server", "validationError" ],
            "nullable" : true
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message.",
            "nullable" : true
          },
          "objects" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named structured values associated with the error."
            },
            "description" : "Named structured values associated with the error."
          },
          "operation" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Operation that failed. Its fields depend on the operation type.",
              "nullable" : true
            },
            "description" : "Operation that failed. Its fields depend on the operation type.",
            "nullable" : true
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named scalar values associated with the error."
            },
            "description" : "Named scalar values associated with the error."
          },
          "status" : {
            "type" : "string",
            "description" : "Symbolic HTTP response status.",
            "enum" : [ "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "PRECONDITION_FAILED", "UNSUPPORTED_MEDIA_TYPE", "UNPROCESSABLE_ENTITY", "PRECONDITION_REQUIRED", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "HTTP_VERSION_NOT_SUPPORTED" ]
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Numeric HTTP response status.",
            "maximum" : 599,
            "minimum" : 400
          },
          "suggestedAction" : {
            "type" : "string",
            "description" : "Recovery action suggested to the client.",
            "enum" : [ "none", "requestRole", "logout", "logoutImmediately", "provideAuthorizationHeader", "refreshToken" ],
            "nullable" : true
          }
        },
        "required" : [ "status", "statusCode" ]
      },
...
      "LogUsageSummary" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "The totals for a range beside the hourly series behind them. The range is echoed back as the\nserver resolved it, so a caller that omitted a bound can see which one it got.\n",
        "properties" : {
          "from" : {
            "type" : "string",
            "description" : "The inclusive lower bound of the range, as a UTC instant."
          },
          "series" : {
            "type" : "array",
            "description" : "One entry per hour that holds a row, oldest first. Hours with no traffic are absent.",
            "items" : {
              "$ref" : "#/components/schemas/TimeBucket"
            }
          },
          "to" : {
            "type" : "string",
            "description" : "The exclusive upper bound of the range, as a UTC instant."
          },
          "totals" : {
            "$ref" : "#/components/schemas/UsageTotals"
          }
        }
      },
...
      "RequestRow" : {
        "type" : "object",
        "properties" : {
          "apiKeyHash" : {
            "type" : "string"
          },
          "authSource" : {
            "type" : "string"
          },
          "component" : {
            "type" : "string"
          },
          "durationNanos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "errorPack" : {
            "type" : "string"
          },
          "globalRequestId" : {
            "type" : "string"
          },
          "handler" : {
            "type" : "string"
          },
          "httpMethod" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "requestTime" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "userId" : {
            "type" : "string"
          }
        }
      },
...
      "ServerBuildReport" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "The version the named server's environment declares, beside the artifact its JVM loaded. A\nserver reporting the release version from an artifact older than the release was pulled and\nnot rebuilt.\n",
        "properties" : {
          "application" : {
            "type" : "string"
          },
          "artifactBuiltAt" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "artifactError" : {
            "type" : "string",
            "description" : "Why the artifact could not be located, where it could not be."
          },
          "artifactPath" : {
            "type" : "string",
            "description" : "Where the server's own code was loaded from. Null where the class loader has no code source.",
            "nullable" : true
          },
          "artifactSizeBytes" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Present for a jar, absent for a class directory."
          },
          "artifactType" : {
            "type" : "string",
            "description" : "Whether that is a packaged jar or a build tree's class directory.",
            "enum" : [ "jar", "classes" ]
          },
          "host" : {
            "type" : "string",
            "nullable" : true
          },
          "implementationVersion" : {
            "type" : "string",
            "description" : "Implementation-Version from the manifest, where the artifact carries one.",
            "nullable" : true
          },
          "java.vendor" : {
            "type" : "string"
          },
          "java.version" : {
            "type" : "string"
          },
          "jvm.name" : {
            "type" : "string"
          },
          "pid" : {
            "type" : "integer",
            "format" : "int64"
          },
          "server" : {
            "type" : "string"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "uptimeMs" : {
            "type" : "integer",
            "format" : "int64"
          },
          "version" : {
            "type" : "string",
            "nullable" : true
          },
          "versionModifier" : {
            "type" : "string",
            "nullable" : true
          }
        }
      },
...
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}